Do not eval unsafe ext_emconf.php
The TER frontend upload needs to extract various information
(description, state, dependecies, title, etc) from the extensions
em_conf.php file.
This file is user generated and could contain malicious code.

As collecting all information via the web interface is not feasible,
we parse the file with PHP-Parser and evalute save parts of the code.

PHP-Parser turns the file into an AST. Then we can traverse it and
remove any nodes that are not scalars or arrays.
Additionally we check that we only have one assignment and that we
only assign to $EM_CONF.

The save AST is then turned back into PHP code and eval'd such that
we get access to the data.

The TER SOAP API is not affected, because the ext_emconf.php file
contents are provided as array and not as file.

Change-Id: I41ac8aedc3a233b5d56caa9c11eed4075d380537
Reviewed-on: http://review.typo3.org/31940
Reviewed-by: Christian Zenker <typo3@xopn.de>
Tested-by: Christian Zenker <typo3@xopn.de>