[WiP][BUGFIX] Don't cache list and listLatest actions
It looks like RealUrl on typo3.org has bugs related to cHash caching.
It happened several times that caching was disabled because the
stored cHash for a speaking URL was wrong.

The fix for this was to clear the tx_realurl_chashcache table.
However, this led to even more problems because now the wrong entries
are displayed on the details page of extensions.

It's assumed that this is all caused by an old (and patched) RealUrl
version on typo3.org. Until it is replaced, the workaround is to change
the list view for extensions to USER_INT.

Change-Id: I407545c319c6248b071b4844aef1d72bedca6d5a
Reviewed-on: https://review.typo3.org/48285
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>