[BUGFIX|TASK] Remove validation and editing for repository clone url
The repository_clone_url of an extension may have a git scheme or
none too, thus the current url validator will fail for those extensions.

Since using the repository_clone_url in composer.typo3.org turned out
to be a bad idea by making things more complicated this functionality
will be removed. As a first step editing the clone url is removed.
Later the functionality from composer.typo3.org will be removed in
favor of supplying it on packagist.

Change-Id: I5a50b873aec9d498ab4a7884dcdd00a5d33947e4
Reviewed-on: https://review.typo3.org/48317
Reviewed-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>